use static identity and tlp:clear on generic extractions #231
